Animated Login New
After re-planning the transitions for the user and the admin pages, respectably, we concluded to the following design.
From the animated Login page the user has the option to sign up straight as a user. Upon logging in, the system will check and determine weather the logged in account is a user or an admin.  Then depending on the response the sites redirects to the equivalent page .